Moved MediaPlayer to Service.
The media player instance has been moved to a service thereby fixing
weird issues where the application would stop playing music when the
device was put to sleep or something else. This also allowed for the
refactoring of a lot of activity dependent code. Notifications for
playback are now created.