More Bug Fixes
Album art is now cached using a LinkedHashMap, audio focus is now
abandoned and the player is stopped when something else requests audio
focus. The app now ducks it’s audio volume as well for transient
losses. Fixed a bug where the queue would not be instantiated in time
for the UI. Also added debug release checks to lock off features of the
app for non debug users.